pyproject.toml
[project]
name = "swine-monitor"
version = "1.0.0"
description = "Swine breeding behavior detection system using YOLO and PyQt6"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=44.0.0",
    "opencv-python>=4.12.0.88",
    "pip>=26.0.1",
    "pyqt6>=6.10.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ultralytics>=8.3.232",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
    "pytest>=7.0.0",
]
